requirements.txt
# Core LLM + RAG
openai>=1.30.0
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-community>=0.2.0
langchain-text-splitters>=0.2.0
chromadb>=0.5.0
tiktoken>=0.7.0

# Document processing
pymupdf>=1.24.0

# Real-time video + YOLO26
ultralytics>=8.4.0
opencv-python-headless>=4.9.0
numpy>=1.26.0

# Flask web framework
flask>=3.0.0
gunicorn>=22.0.0

# Utilities
python-dotenv>=1.0.0
Pillow>=10.0.0
pandas>=2.0.0
